requirements.txt
discord.py==2.3.2
huggingface-hub==0.19.3
requests==2.32.3
gradio==4.44.0
gradio-client==1.3.0
bs4==0.0.2
openai==1.46.0
emoji==2.12.1
logging==0.4.9.6



